ls: Don't open files oustide of root dir
When several folders are added to the workspace the `root` still points
to only one of the folders. This doesn't play nicely with the way
zeta-note expects the notes to be organized and referenced and even lead
to a crash.

This solves the crash, but doesn't solve the use-case with several
folders in the workspace. This needs to be solved separately.