ls: Support multi-folder workspaces
This seems to be mainly a VSCode thing, but nevertheless this is
something I found useful as it allows opening both work notes folder and
personal notes folder in the same VSCode workspace which is much more
convenient than having separate VSCode windows for different note
folders.

The current implementation treats separate folders in isolation.