diag/fix: Re-send diagnostics when the document is reopened
Previously, the logic was simple: if during state update the diagnostics
for a file didn't change there's no need to re-send it.

This is true, when the file is closed/re-opened: no change, no need to
send diagnostics. However, some (all?) editors drop diagnostics when a
file gets closed.

Now the server checks for this case too and re-sends the diagnotics when
needed.

Resolves #62.