[core] Make text sync configurable and use full by default
Incremental text sync is more efficient but has been problematic
historically due to client bugs and protocol lacking necessary
information to enforce stronger validation on the server side.

The option to choose incremental text sync is still there, but the
default now is 'full' where the client is expected to send the full
contents of the document on every change.