[build] Use partial trimming to avoid runtime errors
dotnet 7 turned on partial trimmng by default:
  https://learn.microsoft.com/en-us/dotnet/core/compatibility/deployment/7.0/trim-all-assemblies

This is a breaking change, and although, they say it should
affect only apps with trim warnings, I didn't get any warnings when
building marksman, and still it trimmed some essential methods.

Enabling partial trimming is not a proper fix, but at least it allows to
move to dotnet 7. So, good enough.