core: Unbreak single file mode
In single file mode the initial set of workspace folders may be empty.
Doing minBy on an empty collection raises an exception... which gets
swallowed by dotnet async and effectively causes the server to hang on
init.

This patch does an empty check and also wraps the Initialize method in
try-with (this was the only method of the server that didn't have an
exception handler).