fix: Check for a bogus root folder in workspace/didChangeWorkspaceFolders request (#380)
Sometimes, LSP clients can send bogus folders (CWD, or homedir) as part
of added folders in didChangeWorkspaceFolders. Whenever this happens
marksman treats the root as the real root of the project folders and
recursively scans/indexes mardown documents underneath.This is
problematic when such root is a homedir because it makes marksman scan
too much and potentially fail.

Fixes #377
Potentially, addresses #373 too